Refactor project and seperate the cities from main
